Magic doesn't work

+++ FantasyFPS/Assets/Combat/DealDamage.cs
@@ -5,12 +5,21 @@ using UnityEngine;
 public class DealDamage : MonoBehaviour
 {
     // Start is called before the first frame update
-    public float damage = 10.0f;
+    public int damage = 10;
+    //public Obb obb;
+
 
     void Start()
     {
-        
+        //var collider = this.GetComponent<BoxCollider>();
+        //Obb obb = ObbCollisionDetection.ToObb(collider);
     }
+
+    //public Obb GetObb()
+    //{
+    //    var collider = this.GetComponent<BoxCollider>();
+    //    return ObbCollisionDetection.ToObb(collider);
+    //}
 
     // Update is called once per frame
     void Update()
